webapp/requirements.txt
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
Django==1.8.5
django-bootstrap-form==3.2
django-bower==5.0.4
django-celery==3.1.16
django-filter==0.11.0
django-suit==0.2.15
djangorestframework==3.2.4
gunicorn==19.3.0
kombu==3.0.26
Markdown==2.6.2
meld3==1.0.2
pytz==2015.6
six==1.9.0
supervisor==3.1.3
